refactor: Enhance project structure
