+++ index.html
@@ -1,43 +1,55 @@
 <!DOCTYPE html>
-<meta charset="UTF-8">
-<meta name="viewport" content="width=device-width, initial-scale=1.0">
+<meta charset="UTF-8" />
+<meta name="viewport" content="width=device-width, initial-scale=1.0" />
 <html lang="en">
-<head>
+  <head>
     <title>Découvrez si votre utilisation du numérique pollue !</title>
-    <link rel="stylesheet" href="src/styles/material-theme/css/theme.css">
-    <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,500..700,0..1,-50..200"/>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/jspdf/1.5.3/jspdf.debug.js"></script>   
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/html2canvas/0.5.0-beta4/html2canvas.min.js"></script> 
-    <script src="src/scripts/script.js" defer></script>
-</head>
-<body id="body">
+    <link rel="stylesheet" href="src/styles/material-theme/css/theme.css" />
+    <link
+      rel="stylesheet"
+      href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,500..700,0..1,-50..200"
+    />
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/jspdf/1.5.3/jspdf.debug.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/html2canvas/0.5.0-beta4/html2canvas.min.js"></script>
+    <script src="src/scripts/script.js" defer type="module"></script>
+  </head>
+  <body id="body">
     <header>
-        <button type="button" id="toggleThemeBtn" onclick="toggleTheme();">.</button>
-        <button type="button" id="getImgReferenceBtn"><span class="material-symbols-outlined">image</span></button>
-        <button type="button" id="shareBtn" onclick="copyLink();"><span class="material-symbols-outlined">share</span></button>
+      <button type="button" id="toggleThemeBtn">.</button>
+      <button type="button" id="getImgReferenceBtn">
+        <span class="material-symbols-outlined">image</span>
+      </button>
+      <button type="button" id="shareBtn">
+        <span class="material-symbols-outlined">share</span>
+      </button>
     </header>
     <div id="questionsContainer" class="question-display-box">
-        <div id="progressBarBg" class="progress-bar-bg">
-            <div id="progressBarFg" class="progress-bar-fg"></div>
-        </div>
-        <div class="round-glass" id="divToDelete">
-            <h1 class="on-round-glass">Bienvenue sur EcoQuiz !</h1>
-            <h2 class="on-round-glass">Vous allez effectuer un petit questionnaire pour mieux vous éclairer sur l'impact de votre consommation du numérique.</h2>
-            <h3 class="on-round-glass">Vos réponses ne sont pas sauvegardée. </h3>
-            <h3 class="on-round-glass">Si vous avez du mal à répondre, faites une estimation. </h3>
-            <h3 class="on-round-glass">N'hésitez pas à partager le questionnaire ! </h3>
-            <button id="startSurveyBtn" class="start-btn" onclick="startSurvey();">
-                <span class="button-text">Commencer le questionnaire</span>
-                <span class="material-symbols-outlined">flag</span>
-            </button>
-        </div>
+      <div id="progressBarBg" class="progress-bar-bg">
+        <div id="progressBarFg" class="progress-bar-fg"></div>
+      </div>
+      <div class="round-glass" id="divToDelete">
+        <h1 class="on-round-glass">Bienvenue sur EcoQuiz !</h1>
+        <h2 class="on-round-glass">
+          Vous allez effectuer un petit questionnaire pour mieux vous éclairer
+          sur l'impact de votre consommation du numérique.
+        </h2>
+        <h3 class="on-round-glass">Vos réponses ne sont pas sauvegardée.</h3>
+        <h3 class="on-round-glass">
+          Si vous avez du mal à répondre, faites une estimation.
+        </h3>
+        <h3 class="on-round-glass">
+          N'hésitez pas à partager le questionnaire !
+        </h3>
+        <button id="startSurveyBtn" class="start-btn">
+          <span class="button-text">Commencer le questionnaire</span>
+          <span class="material-symbols-outlined">flag</span>
+        </button>
+      </div>
     </div>
-    <div id="alternatives">
-    </div>
-    <div id="shareDiv">
-    </div>
+    <div id="alternatives"></div>
+    <div id="shareDiv"></div>
     <footer>
-        <img alt="cn2_logo" id="bottomLogo">
+      <img alt="cn2_logo" id="bottomLogo" />
     </footer>
-</body>
+  </body>
 </html>
